agent/pyproject.toml
[tool.poetry]
name = "blog-post-generator"
version = "0.1.0"
description = "A blog post generator"
authors = ["shrey shah <sshreyv@gmail.com>"]
license = "MIT"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
python-dotenv = "*"
langchain-core = "*"
langchain-openai = "*"
langgraph = ">=0.2.39"
langgraph-checkpoint = ">=2.0.2"
langgraph-checkpoint-postgres = ">=2.0.2"
pydantic = "*"
langgraph-cli = {extras = ["inmem"], version = ">=0.1.71"}
fastapi = ">=0.115.8"
copilotkit = ">=0.1.38"

[tool.poetry.scripts]
app = "src.app:main"

[tool.black]
line-length = 220
target-version = ["py39"]
include = '\.pyi?$'

[tool.ruff]
line-length = 220

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
